# api/web/api-doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kinofan API",
        "description": "API для проекта Кинофан",
        "contact": {
            "name": "Nasirov Ismail",
            "email": "isik94@ciloud.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://kinofan.api/v1",
            "description": "API server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "Данный эндпоинт создаeт нового пользователя с ролью user",
                "operationId": "addUser",
                "requestBody": {
                    "description": "Данные для регистрации пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запрос выполнен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пустой или неправильный токен"
                    },
                    "403": {
                        "description": "Запрещено"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "406": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация пользователя",
                "description": "Данный эндпоинт выдает accessToken и refreshToken",
                "operationId": "authUser",
                "requestBody": {
                    "description": "Данные для авторизации пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "username": "isik@yandex.ru",
                                    "password": "qwert123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запрос выполнен успешно"
                    },
                    "401": {
                        "description": "Пустой или неправильный токен"
                    },
                    "403": {
                        "description": "Запрещено"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "406": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Получить новый временный accessToken",
                "description": "Данный эндпоинт выдает refreshToken и новый временный accessToken",
                "operationId": "newRefreshToken",
                "requestBody": {
                    "description": "Данные для получения accessToken",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "refreshToken": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "refreshToken": "dsweERTWEg25GRdghedgaeojkpojkRERGHAABRH"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запрос выполнен успешно"
                    },
                    "401": {
                        "description": "Пустой или неправильный токен"
                    },
                    "403": {
                        "description": "Запрещено"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "406": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Разлогинить пользователя и удалить refreshToken",
                "description": "Данный эндпоинт удаляет refreshToken для пользователя, который вышел из системы",
                "operationId": "deleteRefreshToken",
                "requestBody": {
                    "description": "Данные для получения accessToken",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "refreshToken": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "refreshToken": "dsweERTWEg25GRdghedgaeojkpojkRERGHAABRH"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запрос выполнен успешно"
                    },
                    "401": {
                        "description": "Пустой или неправильный токен"
                    },
                    "403": {
                        "description": "Запрещено"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "406": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "description": "ID пользователя",
                                "type": "integer",
                                "example": "1"
                            },
                            "username": {
                                "description": "Никнейм пользователя",
                                "type": "email",
                                "example": "qwe@gmail.com"
                            },
                            "email": {
                                "description": "Электронная почта пользователя",
                                "type": "email",
                                "example": "qwe@gmail.com"
                            },
                            "status": {
                                "description": "Статус пользователя",
                                "type": "integer",
                                "enum": [
                                    "0",
                                    "9",
                                    "10"
                                ],
                                "example": "10"
                            },
                            "statusText": {
                                "description": "Статус пользователя",
                                "type": "string",
                                "enum": [
                                    "Удален",
                                    "Не активен",
                                    "Активен"
                                ],
                                "example": "Активен"
                            },
                            "created_at": {
                                "description": "Время в unixtime",
                                "type": "integer",
                                "example": "1687464863"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/UserRoles"
                    }
                ]
            },
            "SignupForm": {
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "username": {
                        "description": "Электронная почта пользователя",
                        "type": "string",
                        "example": "isik@yandex.ru"
                    },
                    "password": {
                        "description": "Пароль",
                        "type": "string",
                        "example": "qwerty"
                    }
                },
                "type": "object"
            },
            "UserRoles": {
                "properties": {
                    "userRoles": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "role": {
                                    "description": "Ключ роли пользователя",
                                    "type": "string",
                                    "example": "admin"
                                },
                                "roleTranslate": {
                                    "description": "Перевод роли на русский язык",
                                    "type": "string",
                                    "example": "Администратор"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Методы для работы с пользователями"
        }
    ]
}